// src/apps/McGarden/types/mcPackage.ts
import { McStatus } from './mcStatus';
export type McPackage = {
    siteCode: string;
    projectIdentifier: string;
    projectDescription: string;
    mcPkgId: string;
    mcPkgNumber: string;
    commPkgId: string;
    commPkgNumber: string;
    commPkgStatus: McStatus;
    description: string;
    remark: string;
    system: string;
    subsystem: string;
    area: string;
    discipline: string;
    disciplineDescription: string;
    mcStatus: McStatus;
    tagVolume: string;
    priority: string;
    priorityDescription: string;
    priority2: string;
    priority2Description: string;
    priority3: string;
    priority3Description: string;
    phase: string;
    responsible: string;
    createdDate: string;
    updatedDate: string;
    finalPunchPlannedDate: string;
    finalPunchForecastDate: string;
    finalPunchActualDate: string;
    punchAcceptActualDate: string;
    rfccPlannedDate: string;
    rfccForecastDate: string;
    rfccActualDate: string;
    rfccShippedDate: string;
    tacPlannedDate: string;
    tacForecastDate: string;
    rfocPlannedDate: string;
    rfocForecastDate: string;
    rfocActualDate: string;
    tacIsShipped: boolean;
    tacIsAccepted: boolean;
    isVoided: boolean;
    url: string;
    certificateNumber: string;
    rfocIsShipped: boolean;
    rfocIsAccepted: boolean;
    rfocIsRejected: boolean;
    rfccIsShipped: boolean;
    rfccIsAccepded: boolean;
    rfccIsRejected: boolean;
    searchableValues: string;
};